deploy commit

+++ StudyNotion/Server/controllers/Section.js
@@ -1,21 +1,25 @@
 const Section = require("../models/Section");
 const Course = require("../models/Course");
-
+const SubSection = require("../models/Subsection");
+// CREATE a new section
 exports.createSection = async (req, res) => {
   try {
+    // Extract the required properties from the request body
     const { sectionName, courseId } = req.body;
+
+    // Validate the input
     if (!sectionName || !courseId) {
       return res.status(400).json({
         success: false,
-        message: "Missing Properties",
+        message: "Missing required properties",
       });
     }
 
-    const newSection = await Section.create({
-      sectionName,
-    });
+    // Create a new section with the given name
+    const newSection = await Section.create({ sectionName });
 
-    const updatedCourseDetails = await Course.findByIdAndUpdate(
+    // Add the new section to the course's content array
+    const updatedCourse = await Course.findByIdAndUpdate(
       courseId,
       {
         $push: {
@@ -32,63 +36,97 @@ exports.createSection = async (req, res) => {
       })
       .exec();
 
-    return res.status(200).json({
+    // Return the updated course object in the response
+    res.status(200).json({
       success: true,
       message: "Section created successfully",
-      updatedCourseDetails,
+      updatedCourse,
     });
   } catch (error) {
-    return res.status(500).json({
+    // Handle errors
+    res.status(500).json({
       success: false,
-      message: "Unable to create Section , please try again",
+      message: "Internal server error",
       error: error.message,
     });
   }
 };
 
+// UPDATE a section
 exports.updateSection = async (req, res) => {
   try {
-    const { sectionName, sectionId } = req.body;
-
-    if (!sectionName || !sectionId) {
-      return res.status(400).json({
-        success: false,
-        message: "Missing Properties",
-      });
-    }
-
+    const { sectionName, sectionId, courseId } = req.body;
     const section = await Section.findByIdAndUpdate(
       sectionId,
       { sectionName },
       { new: true }
     );
-
-    return res.status(200).json({
+    const course = await Course.findById(courseId)
+      .populate({
+        path: "courseContent",
+        populate: {
+          path: "subSection",
+        },
+      })
+      .exec();
+    console.log(course);
+    res.status(200).json({
       success: true,
-      message: "Section Updated Successfully",
+      message: section,
+      data: course,
     });
   } catch (error) {
-    return res.status(500).json({
+    console.error("Error updating section:", error);
+    res.status(500).json({
       success: false,
-      message: "Unable to update Section , please try again",
+      message: "Internal server error",
       error: error.message,
     });
   }
 };
 
+// DELETE a section
 exports.deleteSection = async (req, res) => {
   try {
-    const { sectionId } = req.params;
+    const { sectionId, courseId } = req.body;
+    await Course.findByIdAndUpdate(courseId, {
+      $pull: {
+        courseContent: sectionId,
+      },
+    });
+    const section = await Section.findById(sectionId);
+    console.log(sectionId, courseId);
+    if (!section) {
+      return res.status(404).json({
+        success: false,
+        message: "Section not found",
+      });
+    }
+    // Delete the associated subsections
+    await SubSection.deleteMany({ _id: { $in: section.subSection } });
+
     await Section.findByIdAndDelete(sectionId);
 
-    return res.status(200).json({
+    // find the updated course and return it
+    const course = await Course.findById(courseId)
+      .populate({
+        path: "courseContent",
+        populate: {
+          path: "subSection",
+        },
+      })
+      .exec();
+
+    res.status(200).json({
       success: true,
-      message: "Section Deleted Successfully",
+      message: "Section deleted",
+      data: course,
     });
   } catch (error) {
-    return res.status(500).json({
+    console.error("Error deleting section:", error);
+    res.status(500).json({
       success: false,
-      message: "Unable to Delete Section , please try again",
+      message: "Internal server error",
       error: error.message,
     });
   }
